Move ExeTime to separate file
